+++ src/main/java/dk/kb/present/dr/holdback/HoldbackDatePicker.java
@@ -1,4 +1,4 @@
-package dk.kb.present.holdback;
+package dk.kb.present.dr.holdback;
 
 import java.io.FileInputStream;
 import java.io.IOException;
